+++ lib/diffo/validations/is_uuid4_or_nil.ex
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2025 diffo contributors <https://github.com/diffo-dev/diffo/graphs.contributors>
+#
+# SPDX-License-Identifier: MIT
+
 defmodule Diffo.Validations.IsUuid4OrNil do
   @moduledoc """
   Diffo - TMF Service and Resource Management with a difference