+++ test/provider/external_identifier_test.exs
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2025 diffo contributors <https://github.com/diffo-dev/diffo/graphs.contributors>
+#
+# SPDX-License-Identifier: MIT
+
 defmodule Diffo.Provider.ExternalIdentifierTest do
   @moduledoc false
   use ExUnit.Case